update slot history log

+++ models/appointment.go
@@ -4,11 +4,11 @@ package models
 import "time"
 
 type Appointment struct {
-	ID        uint           `gorm:"primaryKey" json:"id"`
+	ID        int64           `gorm:"primaryKey" json:"id"`
 	Version   int            `gorm:"default:0" json:"version"`
-	DoctorID  uint           `json:"doctor_id"`
-	PatientID uint           `json:"patient_id"`
-	SlotID    uint           `gorm:"unique" json:"slot_id"` // ห้ามซ้ำ
+	DoctorID  int64           `json:"doctor_id"`
+	PatientID int64           `json:"patient_id"`
+	SlotID    int64           `gorm:"unique" json:"slot_id"` // ห้ามซ้ำ
 	CreatedAt time.Time      `json:"created_at"`
 
 	Doctor    Doctor         `gorm:"foreignKey:DoctorID"`
